Tidy up codecs
